requirements.txt
# Wake Word
pvporcupine>=3.0.0

# Audio
pyaudio>=0.2.14
webrtcvad>=2.10.0

# STT
openai-whisper>=20231117

# TTS
TTS>=0.22.0

# Utils
PyYAML>=6.0
requests>=2.31.0
numpy>=1.24.0
